requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ckeditor==6.7.3
django-filebrowser==4.0.3
django-grappelli==3.0.10
django-js-asset==3.1.2
django-mptt==0.17.0
django-nested-admin==4.1.4
django-tinymce-4==0.0.27
jsmin==3.0.1
libsass==0.23.0
pillow==11.3.0
pyperclip==1.9.0
python-dotenv==1.1.1
python-monkey-business==1.1.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
Unidecode==1.4.0
